test(channel): split into three modules
